script to get twitter influencer score history
the script currently returns a dictionary with keys being topics and
values being influencers with their score history.
It sometimes breaks on some topic and influencers but works other
times. Not sure it's API issue or not.
The script webping.py takes topic_id_list.txt to get a list of topic_ids